Set no proxy override for MITM proxy
There is a use-case in Platypus integration where the MITM proxy and the
code under test run in the same address space see
https://github.com/zendesk/zis-engine/compare/pravinchandar/zis-1819/integration-testing
for more info. The test code is made to use the MITM proxy via the
environment variable `HTTP_PROXY`, because the test code and MITM run in
the same address space, and MITM is also sourcing proxy from ENV, it
results in a request loop.

To avoid the scenario, I'm adding this hack in zendesk/go-replayers....